+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/transactionstatus/AbstractBakerChangeResult.java
@@ -1,6 +1,7 @@
 package com.concordium.sdk.responses.transactionstatus;
 
 import com.concordium.sdk.crypto.ed25519.ED25519PublicKey;
+import com.concordium.sdk.responses.BakerId;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
@@ -14,6 +15,7 @@ import lombok.experimental.SuperBuilder;
 @SuperBuilder
 @EqualsAndHashCode(callSuper = true)
 public abstract class AbstractBakerChangeResult extends AbstractBakerResult {
+    private final BakerId bakerId;
 
     /**
      * The public VRF key of the baker.